nspawn: Add --syscall-whitelist option
The --syscall-whitelist option allows to exclude user-defined
system calls from the seccomp filter.

We reuse systemd's filter sets to match capabilities and introduce
a new function to seccomp-util to allow adding a filter set but
retaining some of the syscalls that belong to it.

In order to run Docker >1.10.3 inside of system-nspawn containers
the syscall `keyctl` is needed. By default it is included in nspawns
seccomp blacklist.

Fixes #5163

Needed for https://github.com/kinvolk/kubeadm-nspawn/issues/21