cgroup-util: fix assertion if /sys/fs/cgroup is not mounted
After  landing #4670, many tests are now broken when building/running them in a
chroot (mock and schroot alike) where `/sys/fs/cgroup` is not mounted (i. e.
empty):

    <mock-chroot> sh-4.3$ make -j4 && ./test-engine
    Assertion 'cg_update_unified() >= 0' failed at ../src/basic/cgroup-util.c:2341, function cg_unified(). Aborting.

We must be able to use our library and build/run unit tests in an
init-less chroot without cgroups. So instead of crashing on an
assertion, simply return false to show that we are not using the
hybrid/unified cgroup.