mount-util: accept that name_to_handle_at() might fail with EPERM
Container managers frequently block name_to_handle_at(), returning
EACCES or EPERM when this is issued. Accept that, and simply fall back
to to fdinfo-based checks.

Notet that we accept either EACCES or EPERM here, as container managers
can choose the error code and aren't very good on agreeing on just one.

(note that this is a non-issue with nspawn, as we prmit
name_to_handle_at() there, only block open_by_handle_at(), which should
be sufficiently safe).