test: add sd-bus early-fd test case
A dbus connection setup is always special. It uses an ASCII line based
protocol to authenticate via SASL. Once that setup is done, normal DBus
messages are sent. Sadly, the line-based protocol does not allow us to
know message sizes up-front. Hence, we have a buffered reader.

Once the authentication phase is done, we use the trailing buffer data
as input for the first message. From them on, we always read exactly the
required sizes from the wire.

Now if the first message in the trailing SASL buffer contains
file-descriptors, we must be able to properly assign them to the
correct message. This commit adds a test case for exactly that. This
sd-bus does not allow us to have full control over the wire scheduling,
we have to hand-craft some messages here to trigger those code paths.