Never call unmap with MAP_FAILED.
When mmap is called, the code correctly checks for p == MAP_FAILED.

But the resource cleanup at the end of busname_peek_message checks for
p == NULL, and if that's not true, munmap is called.

Therefore in error case, munmap is called with a MAP_FAILED argument
which can result in unexpected behaviour depending on sz's value.

In other files, the pointer is set to NULL in error case, so I go
along with that solution here, too.

Signed-off-by: Tobias Stoeckmann <tobias@stoeckmann.org>